requirements.txt
Flask==3.0.3
flask-cors==4.0.1
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
requests==2.31.0
python-dotenv==1.0.1
